Finished Feather processor and renamed Processors.cs folder
