+++ src/Models/Processors/EggGatherer.cs
